azure-functions/requirements.txt
azure-functions
azure-storage-blob
requests
pymssql